Use Spree::PaymentMethod::CreditCard as base class
`Spree::Gateway` is deprecated and replaced by `Spree::PaymentMethod::CreditCard`.

Since all payment methods inside this gem uses `Spree::CreditCard` as `payment_source_class`
we need to inherit from `Spree::PaymentMethod::CreditCard` instead of the more generic `Spree::PaymentMethod`.

Anyhow, in future versions of solidus we might introduce a minimal source class that does not
plays the role of a credit card. We than can start to separate the classes, but for now
this is ok.